\name{graphics-defunct}
%------ NOTE:  ../R/graphics-defunct.R   must be synchronized with this!
\alias{graphics-defunct}
\alias{GNOME}
\alias{gnome}
\title{Defunct Functions in Package graphics}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
\usage{
gnome(display="", width=7, height=7, pointsize=12)
GNOME(display="", width=7, height=7, pointsize=12)
}
\seealso{
  \code{\link{graphics-deprecated}},
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{internal}
