\name{graphics-package}
\alias{graphics-package}
\alias{graphics}
\docType{package}
\title{
The R Graphics Package
}
\description{
R functions for base graphics
}
\details{
This package contains functions for base graphics.  Base graphics
are traditional S graphics, as opposed 
to the newer \link[grid:grid-package]{grid} graphics.

For a complete
list of functions, use \code{library(help="graphics")}.
}
\author{
R Development Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}

}
\keyword{ package }
