% File src/library/graphics/man/graphics-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{graphics-package}
\alias{graphics-package}
\alias{graphics}
\docType{package}
\title{
The R Graphics Package
}
\description{
R functions for base graphics
}
\details{
  This package contains functions for \sQuote{base} graphics.  Base graphics
  are traditional S-like graphics, as opposed to the more recent
  \link[=grid-package]{grid} graphics.

  For a complete list of functions with individual help pages,
  use \code{library(help = "graphics")}.
}
\author{
  R Core Team and contributors worldwide

  Maintainer: R Core Team \email{R-core@r-project.org}
}
\references{
  Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.

  Murrell, P. (2005) \emph{R Graphics}. Chapman & Hall/CRC Press.
}
\keyword{ package }
